SolarSimulator: onZoom In/Out + CreateMenus
