+++ platform/linux/src/Rtt_LinuxApp.cpp
@@ -59,8 +59,6 @@ namespace Rtt
 	SolarApp::SolarApp()
 		: fSolarGLCanvas(NULL)
 		, fContext(NULL)
-		, fMenuMain(NULL)
-		, fMenuProject(NULL)
 		, fProjectPath("")
 		, fTimer(this, TIMER_ID)
 	{
@@ -100,10 +98,6 @@ namespace Rtt
 	{
 		delete fSolarGLCanvas;
 		delete fContext;
-
-		SetMenuBar(NULL);
-		delete fMenuMain;
-		delete fMenuProject;
 
 		curl_global_cleanup();
 	}
@@ -256,26 +250,6 @@ namespace Rtt
 		GetCanvas()->Refresh(true);
 	}
 
-	void SolarApp::CreateViewAsChildMenu(vector<string>skin, wxMenu* targetMenu)
-	{
-		for (int i = 0; i < skin.size(); i++)
-		{
-			LinuxSimulatorView::SkinProperties sProperties = LinuxSimulatorView::GetSkinProperties(skin[i].c_str());
-			wxMenuItem* currentSkin = targetMenu->Append(sProperties.id, skin[i].c_str(), wxEmptyString, wxITEM_CHECK);
-			Bind(wxEVT_MENU, &SolarApp::OnViewAsChanged, sProperties.id);
-
-			if (sProperties.id == LinuxSimulatorView::Config::skinID)
-			{
-				wxString newWindowTitle(GetContext()->GetTitle());
-				newWindowTitle.append(" - ").append(sProperties.skinTitle.ToStdString());
-
-				LinuxSimulatorView::SelectSkin(sProperties.id);
-				currentSkin->Check(true);
-				SetTitle(newWindowTitle);
-			}
-		}
-	}
-
 	void SolarApp::ClearMenuCheckboxes(wxMenu* menu, wxString currentSkinTitle)
 	{
 		for (int i = 0; i < menu->GetMenuItemCount(); i++)
@@ -285,183 +259,6 @@ namespace Rtt
 			if (!currentItem->GetItemLabel().IsSameAs(currentSkinTitle))
 			{
 				currentItem->Check(false);
-			}
-		}
-	}
-
-	void SolarApp::OnViewAsChanged(wxCommandEvent& event)
-	{
-		int skinID = event.GetId();
-		LinuxSimulatorView::SkinProperties sProperties = LinuxSimulatorView::GetSkinProperties(skinID);
-		SolarApp* frame = solarApp;
-		wxDisplay display(wxDisplay::GetFromWindow(frame));
-		wxRect screen = display.GetClientArea();
-		frame->currentSkinWidth = sProperties.screenWidth;
-		frame->currentSkinHeight = sProperties.screenHeight;
-		int initialWidth = sProperties.screenWidth;
-		int initialHeight = sProperties.screenHeight;
-		wxString newWindowTitle(frame->GetContext()->GetTitle());
-		newWindowTitle.append(" - ").append(sProperties.skinTitle.ToStdString());
-		bool canZoom = sProperties.screenWidth > LinuxSimulatorView::skinMinWidth;
-
-		if (sProperties.selected)
-		{
-			return;
-		}
-
-		frame->fZoomIn->Enable(canZoom);
-		frame->fZoomOut->Enable(canZoom);
-
-		LinuxSimulatorView::Config::skinID = sProperties.id;
-		LinuxSimulatorView::Config::skinWidth = sProperties.screenWidth;
-		LinuxSimulatorView::Config::skinHeight = sProperties.screenHeight;
-		LinuxSimulatorView::SelectSkin(skinID);
-		frame->ClearMenuCheckboxes(frame->fViewAsAndroidMenu, sProperties.skinTitle);
-		frame->ClearMenuCheckboxes(frame->fViewAsIOSMenu, sProperties.skinTitle);
-		frame->ClearMenuCheckboxes(frame->fViewAsTVMenu, sProperties.skinTitle);
-		frame->ClearMenuCheckboxes(frame->fViewAsDesktopMenu, sProperties.skinTitle);
-
-		while (initialWidth > screen.width || initialHeight > screen.height)
-		{
-			initialWidth /= LinuxSimulatorView::skinScaleFactor;
-			initialHeight /= LinuxSimulatorView::skinScaleFactor;
-		}
-
-		LinuxSimulatorView::Config::zoomedWidth = initialWidth;
-		LinuxSimulatorView::Config::zoomedHeight = initialHeight;
-		LinuxSimulatorView::Config::Save();
-
-		frame->GetContext()->SetWidth(initialWidth);
-		frame->GetContext()->SetHeight(initialHeight);
-		frame->ChangeSize(initialWidth, initialHeight);
-
-		wxCommandEvent ev(eventRelaunchProject);
-		wxPostEvent(solarApp, ev);
-	}
-
-	void SolarApp::SetMenu(const char* appPath)
-	{
-		if (LinuxSimulatorView::IsRunningOnSimulator())
-		{
-			const string& appName = GetContext()->GetAppName();
-			SetMenuBar(IsHomeScreen(appName) ? fMenuMain : fMenuProject);
-
-			if (!IsHomeScreen(appName) && fViewMenu->FindItem("View As") == -1)
-			{
-				wxMenu* viewAsMenu = new wxMenu();
-				fViewAsAndroidMenu = new wxMenu();
-				fViewAsIOSMenu = new wxMenu();
-				fViewAsTVMenu = new wxMenu();
-				fViewAsDesktopMenu = new wxMenu();
-				vector<string>namedAndroidSkins;
-				vector<string>genericAndroidSkins;
-				vector<string>namedIOSSkins;
-				vector<string>genericIOSSkins;
-				vector<string>tvSkins;
-				vector<string>desktopSkins;
-				int currentSkinID = ID_MENU_VIEW_AS;
-
-				const char* startupPath = GetStartupPath(NULL);
-				string skinDirPath(startupPath);
-				skinDirPath.append("/Resources");
-				if (!Rtt_IsDirectory(skinDirPath.c_str()))
-				{
-					Rtt_LogException("No Resources dir in %s!\n", startupPath);
-					return;
-				}
-
-				skinDirPath.append("/Skins");
-				wxDir skinDir(skinDirPath);
-				if (!skinDir.IsOpened())
-				{
-					Rtt_LogException("Skin directory not found in /Resources!\n");
-					return;
-				}
-
-
-				if (!skinDir.IsOpened())
-				{
-					Rtt_LogException("Skin directory not found in /Resources!\n");
-					return;
-				}
-
-				wxString filename;
-				lua_State* L = GetContext()->GetRuntime()->VMContext().L();
-				bool fileExists = skinDir.GetFirst(&filename, wxEmptyString, wxDIR_DEFAULT);
-
-				while (fileExists)
-				{
-					if (filename.EndsWith(".lua"))
-					{
-						wxString luaSkinPath(skinDirPath);
-						luaSkinPath.append("/").append(filename);
-
-						LinuxSimulatorView::LoadSkin(L, currentSkinID, luaSkinPath.ToStdString());
-						LinuxSimulatorView::SkinProperties sProperties = LinuxSimulatorView::GetSkinProperties(currentSkinID);
-						wxString skinTitle(sProperties.windowTitleBarName);
-						skinTitle.append(wxString::Format(wxT(" (%ix%i)"), sProperties.screenWidth, sProperties.screenHeight));
-
-						if (sProperties.device.Contains("android") && !sProperties.device.Contains("tv"))
-						{
-							if (sProperties.device.Contains("borderless"))
-							{
-								genericAndroidSkins.push_back(skinTitle.ToStdString());
-							}
-							else
-							{
-								namedAndroidSkins.push_back(skinTitle.ToStdString());
-							}
-						}
-						else if (sProperties.device.Contains("ios"))
-						{
-							if (sProperties.device.Contains("borderless"))
-							{
-								genericIOSSkins.push_back(skinTitle.ToStdString());
-							}
-							else
-							{
-								namedIOSSkins.push_back(skinTitle.ToStdString());
-							}
-						}
-						else if (sProperties.device.Contains("tv"))
-						{
-							tvSkins.push_back(skinTitle.ToStdString());
-						}
-						else if (sProperties.device.Contains("desktop"))
-						{
-							desktopSkins.push_back(skinTitle.ToStdString());
-						}
-					}
-
-					currentSkinID++;
-					fileExists = skinDir.GetNext(&filename);
-				}
-
-				// sort all the skin vectors by name
-				sort(namedAndroidSkins.begin(), namedAndroidSkins.end(), SortVectorByName);
-				sort(genericAndroidSkins.begin(), genericAndroidSkins.end(), SortVectorByName);
-				sort(namedIOSSkins.begin(), namedIOSSkins.end(), SortVectorByName);
-				sort(genericIOSSkins.begin(), genericIOSSkins.end(), SortVectorByName);
-				sort(tvSkins.begin(), tvSkins.end(), SortVectorByName);
-				sort(desktopSkins.begin(), desktopSkins.end(), SortVectorByName);
-
-				// setup the child "view as" menus
-				CreateViewAsChildMenu(namedAndroidSkins, fViewAsAndroidMenu);
-				fViewAsAndroidMenu->AppendSeparator();
-				CreateViewAsChildMenu(genericAndroidSkins, fViewAsAndroidMenu);
-				CreateViewAsChildMenu(namedIOSSkins, fViewAsIOSMenu);
-				fViewAsIOSMenu->AppendSeparator();
-				CreateViewAsChildMenu(genericIOSSkins, fViewAsIOSMenu);
-				CreateViewAsChildMenu(tvSkins, fViewAsTVMenu);
-				CreateViewAsChildMenu(desktopSkins, fViewAsDesktopMenu);
-
-				viewAsMenu->AppendSubMenu(fViewAsAndroidMenu, _T("&Android"));
-				viewAsMenu->AppendSubMenu(fViewAsIOSMenu, _T("&iOS"));
-				viewAsMenu->AppendSubMenu(fViewAsTVMenu, _T("&TV"));
-				viewAsMenu->AppendSubMenu(fViewAsDesktopMenu, _T("&Desktop"));
-				fViewMenu->AppendSubMenu(viewAsMenu, _T("&View As"));
-				fViewMenu->AppendSeparator();
-				fViewMenu->Append(ID_MENU_OPEN_WELCOME_SCREEN, _T("&Welcome Screen"));
 			}
 		}
 	}
@@ -501,85 +298,6 @@ namespace Rtt
 		SetMinClientSize(wxSize(newWidth, newHeight));
 		SetClientSize(wxSize(newWidth, newHeight));
 		SetSize(wxSize(newWidth, newHeight));
-	}
-
-	void SolarApp::OnZoomIn(wxCommandEvent& event)
-	{
-		SolarApp* frame = solarApp;
-		wxDisplay display(wxDisplay::GetFromWindow(frame));
-		wxRect screen = display.GetClientArea();
-		bool doResize = false;
-
-		int proposedWidth = frame->GetContext()->GetWidth() * LinuxSimulatorView::skinScaleFactor;
-		int proposedHeight = frame->GetContext()->GetHeight() * LinuxSimulatorView::skinScaleFactor;
-
-		fZoomOut->Enable(true);
-
-		if (IsHomeScreen(GetContext()->GetAppName()))
-		{
-			doResize = (proposedWidth < screen.width&& proposedHeight < screen.height);
-		}
-		else
-		{
-			if (frame->currentSkinWidth >= proposedWidth && frame->currentSkinHeight >= proposedHeight)
-			{
-				doResize = (proposedWidth < screen.width&& proposedHeight < screen.height);
-			}
-		}
-
-		if (doResize)
-		{
-			frame->GetContext()->SetWidth(proposedWidth);
-			frame->GetContext()->SetHeight(proposedHeight);
-			frame->ChangeSize(proposedWidth, proposedHeight);
-			frame->GetContext()->RestartRenderer();
-			GetCanvas()->Refresh(true);
-
-			if (!IsHomeScreen(GetContext()->GetAppName()))
-			{
-				LinuxSimulatorView::Config::zoomedWidth = proposedWidth;
-				LinuxSimulatorView::Config::zoomedHeight = proposedHeight;
-				LinuxSimulatorView::Config::Save();
-
-				if (proposedWidth * LinuxSimulatorView::skinScaleFactor > screen.width || proposedHeight * LinuxSimulatorView::skinScaleFactor > screen.height)
-				{
-					fZoomIn->Enable(false);
-				}
-			}
-			else
-			{
-				//			LinuxSimulatorView::Config::welcomeScreenZoomedWidth = proposedWidth;
-				//			LinuxSimulatorView::Config::welcomeScreenZoomedHeight = proposedHeight;
-				LinuxSimulatorView::Config::Save();
-			}
-		}
-	}
-
-	void SolarApp::OnZoomOut(wxCommandEvent& event)
-	{
-		SolarApp* frame = solarApp;
-		int proposedWidth = frame->GetContext()->GetWidth() / LinuxSimulatorView::skinScaleFactor;
-		int proposedHeight = frame->GetContext()->GetHeight() / LinuxSimulatorView::skinScaleFactor;
-
-		fZoomIn->Enable(true);
-
-		if (proposedWidth >= LinuxSimulatorView::skinMinWidth)
-		{
-			frame->GetContext()->SetWidth(proposedWidth);
-			frame->GetContext()->SetHeight(proposedHeight);
-			frame->ChangeSize(proposedWidth, proposedHeight);
-			frame->GetContext()->RestartRenderer();
-			GetCanvas()->Refresh(true);
-
-			LinuxSimulatorView::Config::zoomedWidth = proposedWidth;
-			LinuxSimulatorView::Config::zoomedHeight = proposedHeight;
-			LinuxSimulatorView::Config::Save();
-
-			if (proposedWidth / LinuxSimulatorView::skinScaleFactor <= LinuxSimulatorView::skinMinWidth)
-			{
-				fZoomOut->Enable(false);
-			}
-		}
 	}
 
 	void SolarApp::OnSuspendOrResume(wxCommandEvent& event)
@@ -687,103 +405,6 @@ namespace Rtt
 		}
 
 		SetTitle(newWindowTitle);
-	}
-
-	void SolarApp::CreateMenus()
-	{
-		{
-			fMenuMain = new wxMenuBar();
-
-			// file Menu
-			wxMenu* fileMenu = new wxMenu();
-			fileMenu->Append(ID_MENU_NEW_PROJECT, _T("&New Project	\tCtrl-N"));
-			fileMenu->Append(ID_MENU_OPEN_PROJECT, _T("&Open Project	\tCtrl-O"));
-			fileMenu->AppendSeparator();
-			fileMenu->Append(ID_MENU_OPEN_LAST_PROJECT, _T("&Relaunch Last Project	\tCtrl-R"));
-			fileMenu->AppendSeparator();
-			fileMenu->Append(wxID_PREFERENCES, _T("&Preferences..."));
-			fileMenu->AppendSeparator();
-			fileMenu->Append(wxID_EXIT, _T("&Exit"));
-			fMenuMain->Append(fileMenu, _T("&File"));
-
-			// view menu
-			//fViewMenu = new wxMenu();
-			//fZoomIn = fViewMenu->Append(ID_MENU_ZOOM_IN, _T("&Zoom In \tCtrl-KP_ADD"));
-			//fZoomOut = fViewMenu->Append(ID_MENU_ZOOM_OUT, _T("&Zoom Out \tCtrl-KP_Subtract"));
-			//fViewMenu->AppendSeparator();
-			//fMenuMain->Append(fViewMenu, _T("&View"));
-
-			// about menu
-			wxMenu* helpMenu = new wxMenu();
-			helpMenu->Append(ID_MENU_OPEN_DOCUMENTATION, _T("&Online Documentation..."));
-			helpMenu->Append(ID_MENU_OPEN_SAMPLE_CODE, _T("&Sample projects..."));
-			//			helpMenu->Append(ID_MENU_HELP_BUILD_ANDROID, _T("&Building For Android"));
-			helpMenu->Append(wxID_ABOUT, _T("&About Simulator..."));
-			fMenuMain->Append(helpMenu, _T("&Help"));
-		}
-
-		// project's menu
-		{
-			fMenuProject = new wxMenuBar();
-
-			// file Menu
-			wxMenu* fileMenu = new wxMenu();
-			fileMenu->Append(ID_MENU_NEW_PROJECT, _T("&New Project	\tCtrl-N"));
-			fileMenu->Append(ID_MENU_OPEN_PROJECT, _T("&Open Project	\tCtrl-O"));
-			fileMenu->AppendSeparator();
-
-			wxMenu* buildMenu = new wxMenu();
-			buildMenu->Append(ID_MENU_BUILD_ANDROID, _T("Android	\tCtrl-B"));
-			wxMenuItem* buildForWeb = buildMenu->Append(ID_MENU_BUILD_WEB, _T("HTML5	\tCtrl-Shift-Alt-B"));
-			wxMenu* buildForLinuxMenu = new wxMenu();
-			buildForLinuxMenu->Append(ID_MENU_BUILD_LINUX, _T("x64	\tCtrl-Alt-B"));
-			wxMenuItem* buildForARM = buildForLinuxMenu->Append(ID_MENU_BUILD_LINUX, _T("ARM	\tCtrl-Alt-A"));
-			buildMenu->AppendSubMenu(buildForLinuxMenu, _T("&Linux"));
-			fileMenu->AppendSubMenu(buildMenu, _T("&Build"));
-			buildForARM->Enable(false);
-
-			fileMenu->Append(ID_MENU_OPEN_IN_EDITOR, _T("&Open In Editor	\tCtrl-Shift-O"));
-			fileMenu->Append(ID_MENU_SHOW_PROJECT_FILES, _T("&Show Project Files"));
-			fileMenu->Append(ID_MENU_SHOW_PROJECT_SANDBOX, _T("&Show Project Sandbox"));
-			fileMenu->AppendSeparator();
-			fileMenu->Append(ID_MENU_CLEAR_PROJECT_SANDBOX, _T("&Clear Project Sandbox"));
-			fileMenu->AppendSeparator();
-			fileMenu->Append(ID_MENU_RELAUNCH_PROJECT, _T("Relaunch	\tCtrl-R"));
-			fileMenu->Append(ID_MENU_CLOSE_PROJECT, _T("Close Project	\tCtrl-W"));
-			fileMenu->AppendSeparator();
-			fileMenu->Append(wxID_PREFERENCES, _T("&Preferences..."));
-			fileMenu->AppendSeparator();
-			fileMenu->Append(wxID_EXIT, _T("&Exit"));
-			fMenuProject->Append(fileMenu, _T("&File"));
-
-			// hardware menu
-			fHardwareMenu = new wxMenu();
-			wxMenuItem* rotateLeft = fHardwareMenu->Append(wxID_HELP_CONTENTS, _T("&Rotate Left"));
-			wxMenuItem* rotateRight = fHardwareMenu->Append(wxID_HELP_INDEX, _T("&Rotate Right"));
-			//fHardwareMenu->Append(wxID_ABOUT, _T("&Shake"));
-			fHardwareMenu->AppendSeparator();
-			wxMenuItem* back = fHardwareMenu->Append(ID_MENU_BACK_BUTTON, _T("&Back"));
-			fHardwareMenu->AppendSeparator();
-			fHardwareMenu->Append(ID_MENU_SUSPEND, _T("&Suspend	\tCtrl-Down"));
-			fMenuProject->Append(fHardwareMenu, _T("&Hardware"));
-			rotateLeft->Enable(false);
-			rotateRight->Enable(false);
-
-			// view menu
-			fViewMenu = new wxMenu();
-			fZoomIn = fViewMenu->Append(ID_MENU_ZOOM_IN, _T("&Zoom In \tCtrl-KP_ADD"));
-			fZoomOut = fViewMenu->Append(ID_MENU_ZOOM_OUT, _T("&Zoom Out \tCtrl-KP_Subtract"));
-			fViewMenu->AppendSeparator();
-			fMenuProject->Append(fViewMenu, _T("&View"));
-
-			// about menu
-			wxMenu* helpMenu = new wxMenu();
-			helpMenu->Append(ID_MENU_OPEN_DOCUMENTATION, _T("&Online Documentation..."));
-			helpMenu->Append(ID_MENU_OPEN_SAMPLE_CODE, _T("&Sample projects..."));
-			//			helpMenu->Append(ID_MENU_HELP_BUILD_ANDROID, _T("&Building For Android"));
-			helpMenu->Append(wxID_ABOUT, _T("&About Simulator..."));
-			fMenuProject->Append(helpMenu, _T("&Help"));
-		}
 	}
 
 	void SolarApp::StartTimer(float frameDuration)
